package/deb/nm-exp-active-netrics/usr/local/src/nm-exp-active-netrics/src/netrics/plugin/netrics_vca_test/requirements.txt
MouseInfo==0.1.3
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
psutil==5.9.4
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.1
requests==2.25.1
Rx==3.1.1
selenium==3.141.0
six==1.15.0
tinydb==4.4.0
toml==0.10.2
urllib3==1.26.3
